Ignore progress events for ProjectBuilder
Makes tests using ProjectBuilder ignore progress events that are not directly tied to their parent build operations
